chals/web/pasteweb/app/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
requests==2.22.0
urllib3==1.25.3
uuid==1.30
Werkzeug==0.15.5
setuptools==41.0.1
lxml==4.4.1
six==1.12.0
flask-talisman==0.7.0
PyJWT==1.7.1
